start/pause + stop + status & color states
start and pause buttons are now fully operational and tested. Stop button is working but has a bug tied to the state or routing possibly. Once I stop a timer the next timer, does not properly grasp obx bound data. status messages and colors now fully update depending on whether the timer is running or paused.